Fix validation in movie rating, fix add and remove character to movie functionality and fix authentication with username already in use

+++ src/main/java/com/alkemy/disney/entity/MovieEntity.java
@@ -31,7 +31,7 @@ public class MovieEntity {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date releaseDate;
 
-    private float rating;
+    private int rating;
 
     private boolean deleted = Boolean.FALSE;
